+++ source/HTTPServiceSWS.h
@@ -2,6 +2,7 @@
 #define HTTPSERVICESWS_H_
 
 #include <server_http.hpp>
+
 #include "HTTPService.h"
 
 using HttpServer = SimpleWeb::Server<SimpleWeb::HTTP>;
@@ -21,7 +22,17 @@ class HTTPServiceSWS : public HTTPService
 		virtual void startServer();
 
 	private:
-		void defineServerResources(HttpServer& server);
+        void defineServerResources(HttpServer& server);
+
+        void defineServerResourceInfo(HttpServer& server);
+        void defineServerResourceProtocolVersion(HttpServer& server);
+        void defineServerResourceStatus(HttpServer& server);
+        void defineServerResourceBenchResult(HttpServer& server);
+        void defineServerResourcePrepareFile(HttpServer& server);
+        void defineServerResourcePreparePhase(HttpServer& server);
+        void defineServerResourceStartPhase(HttpServer& server);
+        void defineServerResourceInterruptPhase(HttpServer& server);
+        void defineServerResourceErrorHandler(HttpServer& server);
 };
 
 #endif /* HTTPSERVICESWS_H_ */